+++ LiveStream/Base.lproj/Main.storyboard
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<document type="com.apple.InterfaceBuilder3.CocoaTouch.Storyboard.XIB" version="3.0" toolsVersion="14868" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES" initialViewController="BYZ-38-t0r">
+<document type="com.apple.InterfaceBuilder3.CocoaTouch.Storyboard.XIB" version="3.0" toolsVersion="14868" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES" initialViewController="nJI-Rd-xtB">
     <device id="retina6_1" orientation="portrait" appearance="light"/>
     <dependencies>
         <deployment identifier="iOS"/>
@@ -13,18 +13,18 @@
             <objects>
                 <viewController id="BYZ-38-t0r" customClass="ViewController" customModule="LiveStream" customModuleProvider="target" sceneMemberID="viewController">
                     <view key="view" contentMode="scaleToFill" id="8bC-Xf-vdC">
-                        <rect key="frame" x="0.0" y="0.0" width="414" height="896"/>
+                        <rect key="frame" x="0.0" y="0.0" width="414" height="842"/>
                         <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
                         <subviews>
                             <mtkView contentMode="scaleToFill" colorPixelFormat="BGRA8Unorm" depthStencilPixelFormat="Depth32Float" translatesAutoresizingMaskIntoConstraints="NO" id="xCX-qe-mIl" customClass="CanvasMetalView" customModule="LiveStream" customModuleProvider="target">
-                                <rect key="frame" x="0.0" y="44" width="414" height="818"/>
+                                <rect key="frame" x="0.0" y="0.0" width="414" height="808"/>
                                 <gestureRecognizers/>
                                 <connections>
                                     <outletCollection property="gestureRecognizers" destination="uuS-8V-JG2" appends="YES" id="RXY-co-YZm"/>
                                 </connections>
                             </mtkView>
                             <button opaque="NO" contentMode="scaleToFill" contentHorizontalAlignment="center" contentVerticalAlignment="center" buttonType="roundedRect" lineBreakMode="middleTruncation" translatesAutoresizingMaskIntoConstraints="NO" id="WKL-H7-yiK">
-                                <rect key="frame" x="159" y="734" width="96" height="96"/>
+                                <rect key="frame" x="159" y="680" width="96" height="96"/>
                                 <constraints>
                                     <constraint firstAttribute="width" constant="96" id="OcC-PX-FAd"/>
                                     <constraint firstAttribute="height" constant="96" id="Prh-Ek-VPA"/>
@@ -35,7 +35,7 @@
                                 </connections>
                             </button>
                             <switch opaque="NO" contentMode="scaleToFill" horizontalHuggingPriority="750" verticalHuggingPriority="750" contentHorizontalAlignment="center" contentVerticalAlignment="center" on="YES" translatesAutoresizingMaskIntoConstraints="NO" id="O80-Ia-moH">
-                                <rect key="frame" x="306" y="767" width="51" height="31"/>
+                                <rect key="frame" x="306" y="713" width="51" height="31"/>
                                 <constraints>
                                     <constraint firstAttribute="height" constant="31" id="Ciw-KP-ji4"/>
                                     <constraint firstAttribute="width" constant="49" id="GvQ-yy-4dr"/>
@@ -58,8 +58,10 @@
                         </constraints>
                         <viewLayoutGuide key="safeArea" id="6Tk-OE-BBY"/>
                     </view>
+                    <navigationItem key="navigationItem" id="2HV-Oi-c4M"/>
                     <connections>
                         <outlet property="previewView" destination="xCX-qe-mIl" id="xUE-W5-7Fd"/>
+                        <outlet property="recordButton" destination="WKL-H7-yiK" id="naQ-fJ-f0n"/>
                         <outlet property="toggleButton" destination="O80-Ia-moH" id="UCy-kx-y1P"/>
                     </connections>
                 </viewController>
@@ -71,6 +73,34 @@
                 </swipeGestureRecognizer>
             </objects>
             <point key="canvasLocation" x="133" y="123"/>
+        </scene>
+        <!--Start View Controller-->
+        <scene sceneID="XTs-7T-eo0">
+            <objects>
+                <viewController id="nJI-Rd-xtB" customClass="StartViewController" sceneMemberID="viewController">
+                    <view key="view" contentMode="scaleToFill" id="bDq-Rh-jBk">
+                        <rect key="frame" x="0.0" y="0.0" width="414" height="896"/>
+                        <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
+                        <subviews>
+                            <button opaque="NO" contentMode="scaleToFill" contentHorizontalAlignment="center" contentVerticalAlignment="center" buttonType="roundedRect" lineBreakMode="middleTruncation" translatesAutoresizingMaskIntoConstraints="NO" id="ImD-9k-WJ4">
+                                <rect key="frame" x="147.5" y="433" width="119" height="30"/>
+                                <state key="normal" title="MediaLiveStream"/>
+                                <connections>
+                                    <segue destination="BYZ-38-t0r" kind="show" id="MBQ-lS-gja"/>
+                                </connections>
+                            </button>
+                        </subviews>
+                        <color key="backgroundColor" systemColor="systemBackgroundColor" cocoaTouchSystemColor="whiteColor"/>
+                        <constraints>
+                            <constraint firstItem="ImD-9k-WJ4" firstAttribute="centerX" secondItem="bDq-Rh-jBk" secondAttribute="centerX" id="MFa-j4-OgY"/>
+                            <constraint firstItem="ImD-9k-WJ4" firstAttribute="centerY" secondItem="bDq-Rh-jBk" secondAttribute="centerY" id="Rcb-wn-5xM"/>
+                        </constraints>
+                        <viewLayoutGuide key="safeArea" id="rHq-Ep-zoM"/>
+                    </view>
+                </viewController>
+                <placeholder placeholderIdentifier="IBFirstResponder" id="8Zt-M7-YmF" userLabel="First Responder" customClass="UIResponder" sceneMemberID="firstResponder"/>
+            </objects>
+            <point key="canvasLocation" x="-532" y="123"/>
         </scene>
     </scenes>
 </document>
