+++ LiveStream/Info.plist
@@ -2,10 +2,6 @@
 <!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
 <plist version="1.0">
 <dict>
-	<key>NSPhotoLibraryUsageDescription</key>
-	<string>Save video into Photos</string>
-	<key>NSMicrophoneUsageDescription</key>
-	<string>Use microphone to recor video</string>
 	<key>CFBundleDevelopmentRegion</key>
 	<string>$(DEVELOPMENT_LANGUAGE)</string>
 	<key>CFBundleExecutable</key>
@@ -26,6 +22,10 @@
 	<true/>
 	<key>NSCameraUsageDescription</key>
 	<string>This app needs permission to access your cameras</string>
+	<key>NSMicrophoneUsageDescription</key>
+	<string>Use microphone to recor video</string>
+	<key>NSPhotoLibraryUsageDescription</key>
+	<string>Save video into Photos</string>
 	<key>UIApplicationSceneManifest</key>
 	<dict>
 		<key>UIApplicationSupportsMultipleScenes</key>
