+++ LiveStream/StartViewController.swift
@@ -1,0 +1,30 @@
+//
+//  StartViewController.swift
+//  LiveStream
+//
+//  Created by Thang on 4/10/2019.
+//  Copyright © 2019 ThangNVH. All rights reserved.
+//
+
+import UIKit
+
+class StartViewController: UIViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+
+        // Do any additional setup after loading the view.
+    }
+    
+
+    /*
+    // MARK: - Navigation
+
+    // In a storyboard-based application, you will often want to do a little preparation before navigation
+    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
+        // Get the new view controller using segue.destination.
+        // Pass the selected object to the new view controller.
+    }
+    */
+
+}
